構造体の型を変える

+++ src/driver/speaker.h
@@ -7,7 +7,7 @@ namespace driver::speaker
   struct note_t
   {
     uint32_t freq;
-    int64_t ms;
+    int32_t ms;
   };
 
   void init();
